app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=11.1.0",
    "pytest>=8.3.4",
    "python-dotenv>=1.0.1",
    "streamlit>=1.42.2",
    "streamlit-auth>=0.1.7.1",
    "streamlit-extras>=0.5.5",
    "streamlit-oauth>=0.1.14",
    "structlog>=25.1.0",
]
